fix:replace-employees-with-ai-agent-router

+++ app/models/tenant.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, Boolean, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -89,7 +89,8 @@ class PaymentRecord(Base):
     status: Mapped[str] = mapped_column(String(20), default="pending")  # pending, completed, unmatched, failed
     api_key_id: Mapped[str | None] = mapped_column(String(36), nullable=True)
     claimed: Mapped[bool] = mapped_column(Boolean, default=False)  # raw API key has been delivered to the user
-    block_timestamp: Mapped[int] = mapped_column(Integer, default=0)
+    # TronGrid block_timestamp is milliseconds-since-epoch (13 digits) — exceeds int32 range.
+    block_timestamp: Mapped[int] = mapped_column(BigInteger, default=0)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
